requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.8
python-dotenv==1.0.0
django-phonenumber-field==7.3.0
django-filter==23.5
twilio==8.9.1
